Added admin namespace.
