+++ app/controllers/articles_controller.rb
@@ -1,5 +1,5 @@
 class ArticlesController < ApplicationController
-  before_action :set_article, only: [:show, :edit, :update, :destroy]
+  before_action :set_article, only: :show
 
   respond_to :html
 
@@ -12,37 +12,9 @@ class ArticlesController < ApplicationController
     respond_with(@article)
   end
 
-  def new
-    @article = Article.new
-    respond_with(@article)
-  end
-
-  def edit
-  end
-
-  def create
-    @article = Article.new(article_params)
-    @article.save
-    respond_with(@article)
-  end
-
-  def update
-    @article.update(article_params)
-    respond_with(@article)
-  end
-
-  def destroy
-    @article.destroy
-    respond_with(@article)
-  end
-
   private
 
   def set_article
     @article = Article.find(params[:id])
-  end
-
-  def article_params
-    params.require(:article).permit(:title, :content)
   end
 end
