+++ app/views/admin/articles/_form.html.slim
@@ -1,4 +1,4 @@
-= form_for @article do |f|
+= form_for [:admin, @article] do |f|
   - if @article.errors.any?
     #error_explanation
       h2 = "#{pluralize(@article.errors.count, "error")} prohibited this article from being saved:"
