+++ app/views/admin/articles/edit.html.slim
@@ -4,5 +4,4 @@ h1 Editing article
 
 = link_to 'Show', @article
 '|
-= link_to 'Back', articles_path
-
+= link_to 'Back', admin_articles_path
