+++ app/views/admin/articles/index.html.slim
@@ -15,3 +15,9 @@ table
         td = article.title
         td = article.content
         td = link_to 'Show', article
+        td = link_to 'Edit', edit_admin_article_path(article)
+        td = link_to 'Destroy', [:admin, article], data: {:confirm => 'Are you sure?'}, :method => :delete
+
+br
+
+= link_to 'New Article', new_admin_article_path
