+++ app/views/admin/articles/new.html.slim
@@ -1,0 +1,5 @@
+h1 New article
+
+== render 'form'
+
+= link_to 'Back', admin_articles_path
