--- app/views/articles/new.html.slim
@@ -1,5 +1,0 @@
-h1 New article
-
-== render 'form'
-
-= link_to 'Back', articles_path
