+++ app/views/articles/show.html.slim
@@ -7,6 +7,4 @@ p
   strong Content:
   = @article.content
 
-= link_to 'Edit', edit_article_path(@article)
-'|
 = link_to 'Back', articles_path
