+++ test/controllers/articles_controller_test.rb
@@ -11,39 +11,8 @@ class ArticlesControllerTest < ActionController::TestCase
     assert_not_nil assigns(:articles)
   end
 
-  test 'should get new' do
-    get :new
-    assert_response :success
-  end
-
-  test 'should create article' do
-    assert_difference('Article.count') do
-      post :create, article: { content: @article.content, title: @article.title }
-    end
-
-    assert_redirected_to article_path(assigns(:article))
-  end
-
   test 'should show article' do
     get :show, id: @article
     assert_response :success
-  end
-
-  test 'should get edit' do
-    get :edit, id: @article
-    assert_response :success
-  end
-
-  test 'should update article' do
-    patch :update, id: @article, article: { content: @article.content, title: @article.title }
-    assert_redirected_to article_path(assigns(:article))
-  end
-
-  test 'should destroy article' do
-    assert_difference('Article.count', -1) do
-      delete :destroy, id: @article
-    end
-
-    assert_redirected_to articles_path
   end
 end
